+++ fabric/src/main/java/mod/fuji/module/initializer/home/config/model/HomeDataModel.java
@@ -2,6 +2,7 @@ package mod.fuji.module.initializer.home.config.model;
 
 import com.google.common.collect.HashBiMap;
 import com.google.gson.annotations.SerializedName;
+import mod.fuji.core.config.mapper.structure.PlayerKey;
 import mod.fuji.core.structure.GlobalPos;
 import java.util.concurrent.ConcurrentHashMap;
 import lombok.Data;
@@ -13,6 +14,6 @@ import org.jetbrains.annotations.NotNull;
 public class HomeDataModel {
 
     @SerializedName(value = "name2home", alternate = "homes")
-    @NotNull ConcurrentHashMap<String, HashBiMap<String, GlobalPos>> name2home = new ConcurrentHashMap<>();
+    @NotNull ConcurrentHashMap<PlayerKey, HashBiMap<String, GlobalPos>> name2home = new ConcurrentHashMap<>();
 
 }